ai-dashboard/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
librosa>=0.10.0
soundfile>=0.12.0
xgboost>=2.0.0
joblib>=1.3.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
# TensorFlow removed - incompatible with Python 3.14
# The backend will work in fallback/demo mode for deep learning models
